+++ src/openlmis-user/user-profile.routes.js
@@ -60,6 +60,9 @@
                 },
                 roleRightsMap: function(roles, ObjectMapper) {
                     return new ObjectMapper().map(roles, 'rights');
+                },
+                pendingVerificationEmail: function (userId, authUserService) {
+                    return authUserService.getVerificationEmail(userId);
                 }
             }
         });